% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-deps.R
\name{check_deps}
\alias{check_deps}
\title{Check dependency}
\usage{
check_deps(path = ".")
}
\arguments{
\item{path}{path of directory}
}
\description{
Check if the required packages is already installed or not.
}
